refactor: move user funcs from system to services
