+++ nekro_agent/routers/config.py
@@ -16,8 +16,8 @@ from nekro_agent.core.config import (
 from nekro_agent.models.db_user import DBUser
 from nekro_agent.schemas.message import Ret
 from nekro_agent.services.config_service import ConfigService
-from nekro_agent.systems.user.deps import get_current_active_user
-from nekro_agent.systems.user.perm import Role, require_role
+from nekro_agent.services.user.deps import get_current_active_user
+from nekro_agent.services.user.perm import Role, require_role
 from nekro_agent.tools.common_util import get_app_version
