+++ nekro_agent/routers/dashboard.py
@@ -13,7 +13,7 @@ from nekro_agent.models.db_chat_message import DBChatMessage
 from nekro_agent.models.db_exec_code import DBExecCode, ExecStopType
 from nekro_agent.models.db_user import DBUser
 from nekro_agent.schemas.message import Ret
-from nekro_agent.systems.user.deps import get_current_active_user
+from nekro_agent.services.user.deps import get_current_active_user
 
 router = APIRouter(prefix="/dashboard", tags=["Dashboard"])
