+++ nekro_agent/services/agent/templates/j2/system.j2
@@ -109,6 +109,16 @@ Attention: Shared directory is not persistent, it will be cleared after a while.
 * scipy = "^1.14.0"
 * scikit-learn = "^1.5.1"
 * imageio = "^2.35.0"
+* sympy = "^1.13.3"
+* pymupdf = "1.23.8"
+* rarfile = "^4.2"
+* markdown = "^3.7"
+* pandas = "^2.2.3"
+* openpyxl = "^3.1.5"
+* docx = "^0.2.4"
+* python-pptx = "^1.0.2"
+* mplfonts = "^0.0.8"
+* networkx = "^3.4.2"
 
 ### Plugin System & Predefined Method:
